# gdtracker-go-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GDTracker Go API
  version: 0.5.0
  description: |
    Go backend for GDTracker. Session auth and CSRF mirror the Spring API contract (see `docs/AUTH.md`).
    Routes mirror `gdtracker-api` for games, categories, tags, tasks, features, game configuration, archives, game exceptions,
    ingest plane (players, integration ping), game events (definitions + ingest), traces (heatmap), and feedback.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/games/{gameId}/configuration:
    get:
      tags: [Configuration]
      operationId: getGameConfiguration
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          description: OK (bootstraps row if missing)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GameConfiguration" }
        "401": { description: Not authenticated }
        "404": { description: Game not found or not owned }
    patch:
      tags: [Configuration]
      operationId: patchGameConfiguration
      parameters:
        - $ref: "#/components/parameters/GameId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GameConfigurationPatchRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GameConfiguration" }
        "400": { description: Validation error }
        "401": { description: Not authenticated }
        "404": { description: Game not found or not owned }
...
components:
  parameters:
    GameId:
      name: gameId
      in: path
      required: true
      schema: { type: string }
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: ingest-token
  schemas:
...
    CategorySummary:
      type: object
      required: [id, name, color]
      properties:
        id: { type: string }
        name: { type: string }
        color: { type: string }
    ExceptionTaskTemplate:
      type: object
      properties:
        titleTemplate: { type: string }
        descriptionTemplate: { type: string }
        defaultCategoryId: { type: string, nullable: true }
        defaultCategory: { $ref: "#/components/schemas/CategorySummary", nullable: true }
    GameConfiguration:
      type: object
      properties:
        featureFlags:
          type: object
          additionalProperties: { type: boolean }
        settings:
          type: object
          additionalProperties: true
        exceptionTaskTemplate: { $ref: "#/components/schemas/ExceptionTaskTemplate" }
    ExceptionTaskTemplatePatch:
      type: object
      properties:
        titleTemplate: { type: string }
        descriptionTemplate: { type: string }
        defaultCategoryId: { type: string, nullable: true }
    GameConfigurationPatchRequest:
      type: object
      properties:
        featureFlags:
          type: object
          additionalProperties: { type: boolean }
        settings:
          type: object
          additionalProperties: true
        exceptionTaskTemplate: { $ref: "#/components/schemas/ExceptionTaskTemplatePatch", nullable: true }
...